Refactor sso cloud code
